fix/ database.js in wrong location
